Introduce maxSpeechFrames option
This allows for a max number of frames before we "cut" the segment and
continue processing. This is useful for when we want to process a
longer piece of speech (like live speech) and we don't know how large
it will be. That way we have an upper bound on the size of the segment.